♻️ divide main
