+++ src/commands.rs
@@ -1,1 +1,2 @@
+pub mod help;
 pub mod nakachan;
